requirements.txt
asgiref==3.3.0
certifi==2020.6.20
cloudinary==1.22.0
confusable-homoglyphs==3.2.0
Django==3.1.3
django-registration==3.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
